--- src/data/models/SimpleUser.ts
@@ -1,4 +1,0 @@
-export interface SimpleUser {
-  username: string;
-  profilePicture: string;
-}
